backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djoser>=2.2
psycopg2-binary>=2.9
gunicorn>=20.1
python-dotenv>=1.0
django-filter==23.1
drf-extra-fields>=0.7.1
Pillow==11.2.1
flake8==7.2.0